// src/API.ts
export type Post = {
  __typename: "Post",
  id: string,
  message: string,
  owner?: string | null,
  user?: string | null,
  create: string,
  createdAt?: string | null,
  updatedAt: string,
};
// ...
export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items?:  Array<Post | null > | null,
  nextToken?: string | null,
};
// ...
export type CreatePostMutation = {
  createPost?:  {
    __typename: "Post",
    id: string,
    message: string,
    owner?: string | null,
    user?: string | null,
    create: string,
    createdAt?: string | null,
    updatedAt: string,
  } | null,
};
// ...
export type DeletePostMutation = {
  deletePost?:  {
    __typename: "Post",
    id: string,
    message: string,
    owner?: string | null,
    user?: string | null,
    create: string,
    createdAt?: string | null,
    updatedAt: string,
  } | null,
};
// ...
export type ListPostsSortedByCreatedAtQuery = {
  listPostsSortedByCreatedAt?:  {
    __typename: "ModelPostConnection",
    items?:  Array< {
      __typename: "Post",
      id: string,
      message: string,
      owner?: string | null,
      user?: string | null,
      create: string,
      createdAt?: string | null,
      updatedAt: string,
    } | null > | null,
    nextToken?: string | null,
  } | null,
};